adjust regions inside MmioCallbackScope to use u64 for region sizes

+++ bindings/rust/src/lib.rs
@@ -131,7 +131,7 @@ impl Drop for Context {
 }
 
 pub struct MmioCallbackScope<'a> {
-    pub regions: Vec<(u64, usize)>,
+    pub regions: Vec<(u64, u64)>,
     pub read_callback: Option<Box<dyn ffi::IsUcHook<'a> + 'a>>,
     pub write_callback: Option<Box<dyn ffi::IsUcHook<'a> + 'a>>,
 }
